requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
DAWG-Python==0.7.2
docopt==0.6.2
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.0.2
pandas==2.2.3
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyTelegramBotAPI==4.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
